various DFO and AAFC improvements
Modifications to the main program to take care of missing versions in a document, and ensure the first version found result in a document created instead of an add version in case older versions files are missing. This means also that the overall structure of the program swiches from going linear fom file to file to looping inside each set of document and its associated versions.