updates to managing UNIX versions and files for zipping
changes in the converter library to manage UNIX versions properly, and updates in the main program to account for this. Added a couple of utilities to check the files copied vs the files in the xml nodes, and a way to split the uploaded files into folders to zip them in smaller packages easily.